[BREAKING CHANGES] Migrate to Utils
Migrating went well, but it meant that I had to change the self._can_act to await Utils._can_act if we wanted to avoid any attribute errors. So far it looks like everything is working properly as it should be in terms of the intended behavior.
